refactor(structure): moved some files around to reflect the convention a bit better
