--- packages/cli-plugin-metro/src/bundle/index.ts
@@ -1,4 +1,0 @@
-export {default as bundleCommand} from './bundle';
-export {default as ramBundleCommand} from './ramBundle';
-export {buildBundleWithConfig} from './buildBundle';
-export type {CommandLineArgs} from './bundleCommandLineArgs';
