+++ packages/cli-plugin-metro/src/commands/bundle/__mocks__/sign.js
