+++ packages/cli-plugin-metro/src/commands/bundle/__tests__/filterPlatformAssetScales-test.ts
