+++ packages/cli-plugin-metro/src/commands/bundle/__tests__/getAssetDestPathAndroid-test.ts
