+++ packages/cli-plugin-metro/src/commands/bundle/__tests__/getAssetDestPathIOS-test.ts
