+++ packages/cli-plugin-metro/src/commands/bundle/assetPathUtils.ts
