+++ packages/cli-plugin-metro/src/commands/bundle/bundle.ts
