+++ packages/cli-plugin-metro/src/commands/bundle/filterPlatformAssetScales.ts
