+++ packages/cli-plugin-metro/src/commands/bundle/getAssetDestPathAndroid.ts
