+++ packages/cli-plugin-metro/src/commands/bundle/getAssetDestPathIOS.ts
