+++ packages/cli-plugin-metro/src/commands/bundle/index.ts
@@ -1,0 +1,2 @@
+export {default as bundleCommand} from './bundle';
+export {default as ramBundleCommand} from './ramBundle';
