+++ packages/cli-plugin-metro/src/commands/bundle/saveAssets.ts
