+++ packages/cli-plugin-metro/src/commands/index.ts
@@ -1,0 +1,4 @@
+import {bundleCommand, ramBundleCommand} from './bundle';
+import startCommand from './start';
+  
+export default [bundleCommand, ramBundleCommand, startCommand];
