+++ packages/cli-plugin-metro/src/commands/start/watchMode.ts
