+++ packages/cli-plugin-metro/src/index.ts
@@ -1,13 +1,8 @@
 export {
-  bundleCommand,
-  ramBundleCommand,
-  buildBundleWithConfig,
-  CommandLineArgs,
-} from './bundle';
-export {
   Config,
   ConfigLoadingContext,
   MetroConfig,
   getDefaultConfig,
-} from './config';
-export {default as startCommand} from './start';
+  default as loadMetroConfig,
+} from './tools/loadMetroConfig';
+export {default as commands} from './commands';
