+++ packages/cli-plugin-metro/src/tools/__tests__/loadMetroConfig-test.ts
@@ -1,4 +1,4 @@
-import {getDefaultConfig} from '../config';
+import {getDefaultConfig} from '../loadMetroConfig';
 
 jest.mock('fs');
 jest.mock('path');
