+++ packages/cli-plugin-metro/src/tools/metroPlatformResolver.ts
