+++ packages/cli/src/commands/index.ts
@@ -1,9 +1,5 @@
 import {Command, DetachedCommand} from '@react-native-community/cli-types';
-import {
-  startCommand as start,
-  bundleCommand as bundle,
-  ramBundleCommand as ramBundle,
-} from '@react-native-community/cli-plugin-metro';
+import {commands as metroCommands} from '@react-native-community/cli-plugin-metro';
 import link from './link/link';
 import unlink from './link/unlink';
 import install from './install/install';
@@ -16,9 +12,7 @@ import doctor from './doctor';
 import profileHermes from '@react-native-community/cli-hermes';
 
 export const projectCommands = [
-  start,
-  bundle,
-  ramBundle,
+  ...metroCommands,
   link,
   unlink,
   install,
